[1] Confirmation before report comment, delete comment, delete post feature added BE/FE in Javascript (update files with the feature : postView.php, listPostsView.php, adminView.php

+++ view/backend/adminView.php
@@ -23,10 +23,9 @@ $id = (int) $datasAdmin['id'];
             <?= substr($datasAdmin['content'], 0, 600) . "..." ?>
                 <button><a href="index.php?action=post&id=<?=$id?>">Lire la suite</a></button><br/>
                 <button><a href="index.php?action=post&id=<?=$id?>#commentsAnchor">Commentaires</a></button>
-
         </p>
         <button class="adminBtns">Modifier</button>
-        <button class="adminBtns"><a href="index.php?action=deletePost&amp;id=<?= $id?>">Supprimer</a></button>
+        <button class="adminBtns"><a href="index.php?action=deletePost&amp;id=<?= $id?>" onclick="return confirm('Etes vous sûr?')">Supprimer</a></button>
     </div>
 <?php
 }
