+++ view/frontend/listPostsView.php
@@ -19,7 +19,7 @@ $id = (int) $datas['id'];
 
         </p>
         <button class="adminBtns">Modifier</button>
-        <button class="adminBtns"><a href="index.php?action=deletePost&amp;id=<?= $id?>">Supprimer</a></button>
+        <button class="adminBtns"><a href="index.php?action=deletePost&amp;id=<?= $id?>" onclick="return confirm('Etes vous sûr?')">Supprimer</a></button>
     </div>
 <?php
 }
